Delete POSIX target
All functionality has been converted to use Foundation or moved into
Basic (where it can be slowly converted).

<rdar://problem/49197614>  Remove POSIX module from SwiftPM
