+++ Sources/Basic/README.md
@@ -1,7 +1,7 @@
 # Basic Library
 
 This library contains the basic support facilities for `swiftpm`. It is layered
-on `SPMLibc`, `POSIX`, and `Foundation`, and defines the shared data structures,
+on `SPMLibc`  and `Foundation`, and defines the shared data structures,
 utilities, and algorithms for the rest of the package manager.
 
 This library is also intended to contain the bulk of the cross-platform
