--- Tests/POSIXTests/XCTestManifests.swift
@@ -1,1 +1,0 @@
-
